updated clean

+++ bin/clean.sh
@@ -9,9 +9,15 @@ input_folder='../inputs'
 
 for input in ${input_folder}/*.txt.*
 do
-  echo == removing file ${input} ==
-  rm ${input}
+  # test for empty case
+  if [ "$input" != "../inputs/*.txt.*" ]; then
+    echo == removing file ${input} ==
+    rm ${input}
+  fi
 done
 
 rm -f vgcore*
 
+make -C ../CircuitRouter-ParSolver clean
+make -C ../CircuitRouter-SeqSolver clean
+make -C ../lib clean
